# [improvement](join) Avoid copying column elements when join build ouput block